update actions regarding deployment
delete Snapshot action, Snapshots AND normal versions will automatically be build on push to master (according to the version in the pom.xml). Beforehand, this was implicitly done as well, but usually caused one of these actions to fail, because it cannot republish a non-Snapshot version more than once. However, because you can push the tag after the commit, this could happen.

verify-action is only triggered on branches besides the master branch. Pushing on master branch automatically triggers the deploy action that also verifies the build.