update URI to OrderedListOntology
The underlying ontology is the same, but the URI was updated to use the version hosted on github. This has many advantages: We can adapt the ontology if needed and we are not as often rate limited because github seems to be more generous in comparison to the webhoster of purl etc.