remove SparqlOntologyConnector
The performance did not increase that much. The key is to not query the ontology too often with too expensive queries. Queries that need to check labels or localnames are very expensive. These need to be reduced as much as possible. Easiest way is to cache these as we usually expect most of these (e.g., for properties or classes) to not change.