add corefences to corresponding noun mappings
adds the found coreferences to the noun mappings by checking if one word in the cluster belongs to one or more noun mappings. Then adds the coreference to the found noun mapping(s). Currently, they are filtered and we only add Pronoun-related coreferences. These contain less false positives and do not torpedo to much the precision