add merging of nounMappings for Phrases
Before a new NounMapping is added, check first if there is a similar NounMapping. If so, then add the phrase related stuff to the existing NounMapping.
Additionally, added a boolean to NounMappings to show that a NounMapping contains or is a phrase. The intention is that Phrases have a higher likelyhood to be RecommendedInstances