add direct comparison of NounMappings
In TextState, removed the HashMap that mapped a reference to a NounMapping. The reference was used to cluster similar NounMappings.
The similarity-comparison was moved to SimilarityUtils. Important there is the fact that words/references with a separator (e.g., "test.util") are also split and the first part is checked for similarity.