cleanup inconsistency-module (new package, remove commented code)
create new package for types of Inconsistencies for better overview.
remove commented code that was about removing candidates for which the words are already covered by other RecommendedInstances that have a trace link. Maybe revert that at a later point.
Also, cleanup of code and remove need for class variable "candidates" in MissingModelElementInconsistencyAgent