Remove stem and lowercase option from WordVectorSqliteImporter, add filter
Instead of enforcing which type of stemming algorithm is used, this leaves the choice of stemmer to the class extending this base importer.
Additionally, remove lowercasing option since this can also be realized through extending the process() method.
Also add a filter() method which can be extended to skip certain words.