Transform TextExtraction and related classes to new pipeline format
!! DOES NOT COMPILE
Important: Redid most of the API-Stuff. There, we also need much more changes and rethinking of needed/unneeded abstract classes and interfaces
Also, Data that is put into the DataRepository should best introduce a public string with their ID, so they know how to retrieve. This is compliant with information hiding principle because the using classes know they want to use that class either way