Feature: Add ConfigurationHelper and Builder for ArDoCoRunner (#189)
This aims to improve the ArDoCoRunner to be able to provide additional user interfaces more easily.
Specifically, it changes the following:

- Adding a builder to ArDoCoRunner instead of relying on the constructor. This also allows the caller to have String arguments for Files instead of needing to create the File themselves.
- Adding a ConfigurationHelper that can provide all configuration options as well as their default values. Additionally, moved the loadAdditionalConfigs to this helper class

Originally, we wanted to update the text provider CoreNLPProvider as well by allowing to start a CoreNLPServer that loads the models in the background. Unfortunately, there are issues with the CoreNLPServer that we could not solve (yet). Apparently, the state of the CoreDocument (precisely the CoreLabels) is different because some parts seem to not be properly initialized after being transfered. We guess it is due to the fact that serialization is not done properly. Moreover, we figured out that using a server only locally does not bring many benefits as the models are cached within the same JVM/run. Therefore, we would only need the server if we want to deploy the CoreNLPServer elsewhere. This decision will be postponed.