Remove empty lines in gold standard loading (#282)
Due to a minor bug, empty lines in the CSVs of goldstandards were not properly omitted. As such, the goldstandards (in some cases like historical TeaStore UME) expected more than intended because of the empty line (that naturally was never in the results. This bug is fixed now by removing blank lines when loading goldstandard files.